feat: Update Matplotlib plugin to register widget view
- Now can be embedded and/or used in deephaven.ui
- Renamed components from MatPlotLib to Matplotlib to match the product name
- Styling set directly on element... can't register a styles component, should we have an app level render for components?
- Tested with deephaven.ui:
```
from deephaven import ui

@ui.component
def foo():
    import matplotlib.pyplot as plt
    import numpy as np

    from matplotlib.patches import PathPatch
    from matplotlib.path import Path

    N = 400
    t = np.linspace(0, 2 * np.pi, N)
    r = 0.5 + np.cos(t)
    x, y = r * np.cos(t), r * np.sin(t)

    fig, ax = plt.subplots()
    ax.plot(x, y, "k")
    ax.set(aspect=1)
    return fig

f = foo()
```
