--- plugins/matplotlib/src/js/src/Matplotlib.scss
@@ -1,13 +1,0 @@
-@import '@deephaven/components/scss/custom.scss';
-
-.matplotlib-view {
-  height: 100%;
-  width: 100%;
-  display: contents;
-
-  img {
-    height: 100%;
-    width: 100%;
-    object-fit: contain;
-  }
-}
