+++ plugins/matplotlib/src/js/src/index.ts
@@ -1,2 +1,6 @@
+import { MatplotlibPlugin } from './MatplotlibPlugin';
+
 export * from './MatplotlibPanel';
-export * from './DashboardPlugin';
+export * from './MatplotlibView';
+
+export default MatplotlibPlugin;
